Proof of concept: 'numbers' completer
We add the core traits for Completion and Completer (i.e., completion
provider). There will be multiple completion providers defined in the
project, which the user will be able to select dynamically.

We also implement the basic UI for selecting one of the completions by
using arrow keys in the terminal, and for typing the query.

The core traits are implemented by a simple proof-of-concept 'numbers'
completion provider which offers completions in the form of numbers
from 0 up to a given number.

Searching is not implemented yet.