Pass completions in Rc instead of Box
We change the wrapper class in which completions are passed from
completer objects to the UI from Box to Rc. This is a step in
preparation to implementing filtering, where completions which are not
currently shown because of the filter query will have to be retained
in the completer so that they can be recalled quickly when the query
is changed or deleted.