Fs: Recurse into directories to find completions
We modify the Filesystem completer so that it recurses into
directories using a breadth-first search algorithm. When the completer
is pointed at a path, it provides completions coming not only from the
directory's children, but also their children, etc.

This is a step in preparation for implementing search in the
Filesystem completer, where we will want to search not only for
entries directly within the visited directory, but also deeper into
the directory hierarchy.