Fs: Fetch completions asynchronously
We modify the Filesystem completer so that it only yields the contents
of the current directory at first and fetches the rest of potential
completions in a separate thread of execution. This will enable the
completer to quickly yield some output and search for completions when
a search query is entered.