Ui: Display new completions on each key press
When a completer fetches its completions asynchronously, we have to
update the view to show them as they become available. At this point,
this 'refresh' operation is only going to take place after handling a
key press from the user. A better mechanism needs to be developed so
that the entries can appear more smoothly.