Ui: Improve displaying newly fetched completions
We modify the process of displaying completions so that the view is
periodically refreshed if no input is received to include the newly
fetched completions. To this end, the UI module has to spawn a thread
which reads input keys and sends them to the main UI thread, so that
the recv_timeout() method may be used to achieve this.