Fs, Core: Add implementation of text search
We implement searching within the completions found by the Filesystem
completer. To this end, we change the return type of
Completer::completions(): it turned out that we have to return
Arc<Completion> for easy implementation of the Filesystem
completer. The completer had to store the collection of all
completions and the collection of filtered completions separately. To
avoid cloning completions, we turned to Arc for storing completions.

In this commit, we also remove the proof-of-concept Numbers completer;
it was becoming a burden to maintain it.