init: Allow to specify build type in init.sh
We modify init.sh to accept the build type ('debug' or 'release') as
its parameter to allow testing the release version to benchmark
performance. This is for development purposes only; the script will
have to be modified to include a more robust path to the completers
binary before any release can take place.