Fix descending into subdirectories
We fix a bug with descending into subdirectories where only the last
path element of a found completion was taken into consideration when
constructing the path for the descended completer, even when there
were multiple path elements. This led to the creation of completers
for non-existent directories, with empty search results.