Fix descending into directories
We further fix the behavior of descend() for the filesystem completer,
which suffered from an issue where a completer for a non-existent
directory could be created, e.g., when first ascending to a parent
directory and then performing a search. This was caused by the fact
that the path of the selected completion was joined with the path of
the current completer, which was wrong - the proper behavior is to
always take just the path from the selected completer, because it is
always either absolute or relative to the directory from which the
completers binary was run.