Filter completions by subsequence
As the first step to the actual fuzzy match algorithm, we implement a
filter on the completions which only preserves those completions which
contain the query as a subsequence. This may serve as a fuzzy match in
and of itself, but it will also be useful in the future: the fuzzy
match algorithm will rely on a dynamic programming approach, so the
performance will benefit if we can limit the number of completions
which have to go through the scoring phase.