Implement fuzzy match algorithm
We enhance the 'scoring' module by implementing a fuzzy match
algorithm. The algorithm relies on a dynamic programming approach to
compute a "score" for each candidate string against the given query.

The algorithm is not used yet; we need to refactor the "core" code to
make room for keeping the scores and sorting the completions based on
the scores.