Refactor: Filter on higher level than completers
As the next step towards fuzzy matching, we refactor the code so that
the query is no longer passed into the individual completers. Instead,
we filter the completions on a higher level, common for all
completers.

To make this possible, we extend the trait `core::Completion` to
include a new routine `search_string()` which returns the string which
should be subject to filtering and scoring with the search.