Refactor to only compute scores once
We refactor how completions are stored so that we can store the
computed scores along with them; this allows us to run the scoring
algorithm once for each completion.

The change entails moving the responsibility for storing completions
from the individual completers to the `CompleterView` structure. This
is beneficial because the code becomes less repetitive; we can now run
the scoring algorithm in only one place, and store the results so that
the algorithm is only run once for each completion.

Instead of a `completions()` method, completers now have a
`fetch_completions()` method to hand the newly fetched completions
over to the "model" structure `CompleterView`, closer to the UI
module.