Refactor: Avoid storing copies of completion pointers
We refactor the code for associating scores with completions in
'model.rs': instead of a copy of the completion pointer, we store an
index into the collection of all available completions. This lets us
avoid copying Arc pointers, which is a small performance gain, but it
is also a step towards not needing reference-counted pointers to
completions at all, which will serve to simplify the data structures
used in the whole program.

We also replace the 'completions()' method which used to needlessly
clone a large vector of completions-with-scores with a new method
'completion_at(index)', which returns a reference to a single
completion. This has a noticeable performance gain: with large
directory trees, we are no longer seeing a slowdown towards the end of
gathering the completions, and moving around the list feels more
slick.