Optimize: Use array2d instead of nested vectors to reduce allocations
In the dynamic scoring algorithm, we need to create a two-dimensional
array. For longer queries, relying on vectors of vectors is
ineffective, because we need many heap allocations to get the full
array. We use a dedicated external crate 'array2d' to reduce the
number of heap allocations needed to create the array.