Optimize: Do not sort directory entries
We remove the unnecessary step of sorting the completions from a
single directory by their "result string". Now that we have a proper
search mechanism, users will probably rely more on that than on the
entries being sorted in any particular way. This change significantly
reduces the time to "load" the completions from a large directory
tree.